--- packages/components/asset/src/Asset.styles.ts
@@ -1,8 +1,0 @@
-import { css } from 'emotion';
-import tokens from '@contentful/f36-tokens';
-
-export const styles = {
-  asset: css({
-    color: tokens.colorRedBase,
-  }),
-};
