+++ packages/components/asset/src/AssetIcon/AssetIcon.css
