+++ packages/components/asset/src/AssetIcon/AssetIcon.tsx
