+++ packages/components/asset/src/index.ts
@@ -1,2 +1,4 @@
-export { Asset } from './Asset';
-export type { AssetProps } from './Asset';
+export { Asset, isAssetType } from './Asset';
+export type { AssetProps, AssetStatus, AssetType } from './Asset';
+export { AssetIcon } from './AssetIcon/AssetIcon';
+export type { AssetIconProps } from './AssetIcon/AssetIcon';
