--- packages/forma-36-react-components/src/components/Asset/index.ts
@@ -1,4 +1,0 @@
-export { Asset, isAssetType } from './Asset';
-export type { AssetProps, AssetStatus, AssetType } from './Asset';
-export { AssetIcon } from './AssetIcon/AssetIcon';
-export type { AssetIconProps } from './AssetIcon/AssetIcon';
